aux_requirements.txt
python-dotenv==1.0.0
azure-cognitiveservices-speech==1.34.1
